+++ io/src/main/scala/fs2/io/net/tcp/Socket.scala
@@ -21,6 +21,7 @@
 
 package fs2
 package io
+package net
 package tcp
 
 import scala.concurrent.duration._
