+++ io/src/main/scala/fs2/io/net/tls/InputOutputBuffer.scala
@@ -21,6 +21,7 @@
 
 package fs2
 package io
+package net
 package tls
 
 import java.nio.{Buffer, ByteBuffer}
