+++ io/src/main/scala/fs2/io/net/udp/AsynchronousSocketGroup.scala
@@ -21,6 +21,7 @@
 
 package fs2
 package io
+package net
 package udp
 
 import scala.collection.mutable.PriorityQueue
