+++ io/src/main/scala/fs2/io/net/udp/udp.scala
@@ -21,6 +21,7 @@
 
 package fs2
 package io
+package net
 
 /** Provides support for UDP networking. */
 package object udp
