+++ io/src/test/scala/fs2/io/net/tls/TLSDebugExample.scala
@@ -21,6 +21,7 @@
 
 package fs2
 package io
+package net
 package tls
 
 import javax.net.ssl.SNIHostName
